requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
Django==4.2.4
django-baton==2.8.0
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysqlclient==2.2.0
Pillow==10.0.0
pycparser==2.21
pyOpenSSL==23.2.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
